% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{apply.nber.removals}
\alias{apply.nber.removals}
\title{it assumes that there is no removals for the first word}
\usage{
apply.nber.removals(org.names)
}
\description{
it assumes that there is no removals for the first word
}
