% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{browse_dot_graph}
\alias{browse_dot_graph}
\title{Generates a temporary html file with visualization of given nstandr procedures and opens it in a browser (specified in \code{options('browser')}). This requires \code{DiagrammeR} package to be installed. If you do not have those you can \code{cat()} the returned string from `nstandr:::make_dot_graph()' to the R console and copy it to some web tool for dot visualization (e.g.http://magjac.com/graphviz-visual-editor)}
\usage{
browse_dot_graph(procedures_list, width = NULL, height = NULL, ...)
}
\arguments{
\item{procedures_list}{List of procedures. See \code{sandardize} for how to specify of procedures list.}

\item{width}{width of the procedures graph in pixels}

\item{height}{height of the procedures graph in pixels}

\item{...}{
  Arguments passed on to \code{\link[=make_dot_graph]{make_dot_graph}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An object of class ‘htmlwidget’ that can be used in Rmarkdown
}
\description{
Generates a temporary html file with visualization of given nstandr procedures and opens it in a browser (specified in \code{options('browser')}). This requires \code{DiagrammeR} package to be installed. If you do not have those you can \code{cat()} the returned string from `nstandr:::make_dot_graph()' to the R console and copy it to some web tool for dot visualization (e.g.http://magjac.com/graphviz-visual-editor)
}
