% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_harmonize_options.r
\name{check_rows}
\alias{check_rows}
\title{Assumes that rows (if logical) are same length as x}
\usage{
check_rows(
  rows,
  x,
  null_ok = TRUE,
  na_ok = FALSE,
  report_for_call = sys.call(which = -1)
)
}
\arguments{
\item{rows}{either numeric of logical vector}

\item{x}{The object to harmonize}

\item{null_ok}{Whether NULL is valid value}

\item{na_ok}{Whether NA is valid value}

\item{report_for_call}{When reporting issues which function call to indicate for reference}
}
\value{
nothing
}
\description{
Assumes that rows (if logical) are same length as x
}
