% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cockburn.r
\name{cockburn.detect.inst.german}
\alias{cockburn.detect.inst.german}
\title{Detects German Non-profit institutes}
\usage{
cockburn.detect.inst.german(
  x,
  codes.name = "entity.type",
  x.codes.merge = TRUE,
  ...
)
}
\arguments{
\item{x}{table}

\item{...}{
  Arguments passed on to \code{\link[=harmonize.detect]{harmonize.detect}}
  \describe{
    \item{\code{patterns}}{Accepts both vector or table. If patterns it is table can also include replacements column.}
    \item{\code{patterns.col}}{If patterns is not a vector specifies which column to use. Default is 1.}
    \item{\code{patterns.as.codes}}{If set then use patterns as codes. In that case codes are ignored and there is no need to provide them. Default is FALSE.}
    \item{\code{patterns.codes.col}}{If patterns is table which column to use as codes column. Default is 2.}
    \item{\code{patterns.type}}{Specifies kind(s) of patterns. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling code{\link[stringi]{stri_replace_all_regex}}). Another possible options are "exact" and "trim.exact" which matches a pattern string exactly or its trimmed version (i.e., white spaces are ignored at the beginning and at the end of the pattern)}
    \item{\code{patterns.type.col}}{Column in patterns table where you can specify types of patterns for each pattern. If set then \code{patterns.type} is ignored. Default is NULL}
    \item{\code{codes}}{If provided use it as codes. Should be the same length as patterns. Default is NULL.}
    \item{\code{codes.name}}{If provided use it as a name for codes column in results. It also will work as \code{x.codes.col} if \code{x.codes.merge} or \code{x.codes.update.empyty} are set. In case if \code{x.codes.merge} or \code{x.codes.update.empyty} are set and names(\code{x}) does not have column name \code{codes.name} it will reset both \code{x.codes.merge} and \code{x.codes.update.empyty} to FALSE and will (pre)append a new column \code{codes.name}.}
    \item{\code{codes.suffix}}{If \code{codes.name} is not provided use this suffix to append to \code{x.col} name or \code{x.atomic.name} if x is vector (see harmonize.x function).}
    \item{\code{codes.omitted.val}}{If \code{x.rows} is set. Use this value to fill the omitted rows. Default is vector of NAs of \code{x} length. When we update existing codes column (i.e., \code{x.codes.col} is set and valid) and the value is not set explicitly it resets to NULL which means that we use initial values for omitted rows.}
    \item{\code{codes.prepend}}{Whether to prepend codes column to the \code{x}. Default is FALSE. If set it will prepend codes column to \code{x} instead of append.}
    \item{\code{x.codes.col}}{If \code{x} is table, which column to use for making/merging/adding newly detected codes. Default is NULL. When it is not set explicitly and \code{x.codes.merge} or \code{x.codes.update.empty} are set it will first try to use \code{codes.name} if it is provided or if not will set the last column of \code{x} as \code{x.codes.col}.}
    \item{\code{x.codes.update.empty}}{If set then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).}
    \item{\code{x.codes.merge}}{If set then merge (append or prepend) new codes to existing ones. It might turn the codes column to list of vectors.}
    \item{\code{x.codes.merge.prepend}}{If set then it will prepend new codes if not then it will append (default). It is ignored if \code{x.codes.merge} is not set.}
    \item{\code{return.only.codes}}{If set then just return codes vector. Default is FALSE. Basically when it is set it changes \code{return.x.cols} to 0.}
    \item{\code{return.only.first.detected.code}}{If TRUE then return only codes for the first detected pattern. If FALSE return list of vectors of all matched codes. Default is TRUE.}
  }}
}
\value{
Harmonized table
}
\description{
"EINGETRAGENER VEREIN. NON PROFIT SOCIETY/ASSOCIATION."
}
