% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cockburn.r
\name{cockburn_detect_inst_german}
\alias{cockburn_detect_inst_german}
\title{Detects German Non-profit institutes}
\usage{
cockburn_detect_inst_german(
  x,
  output_codes_col_name = "entity_type",
  merge_existing_codes = "append_to_existing",
  ...
)
}
\arguments{
\item{x}{table}

\item{output_codes_col_name}{same as in \code{\link[=detect_patterns]{detect_patterns()}}}

\item{merge_existing_codes}{same as in \code{\link[=detect_patterns]{detect_patterns()}}}

\item{...}{
  Arguments passed on to \code{\link[=harmonize_options]{harmonize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to harmonize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (harmonized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{x_atomic_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_col_name}}{Use this name for the column with results (harmonized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}
}
\value{
Harmonized table
}
\description{
"EINGETRAGENER VEREIN. NON PROFIT SOCIETY/ASSOCIATION."
}
