% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_defactor.r
\name{defactor}
\alias{defactor}
\title{Defactor the object}
\usage{
defactor(
  x,
  conv2dt = c("only.tables", "all.but.atomic", "all.but.lists", "all", "none"),
  x_atomic_name = NULL,
  ...
)
}
\arguments{
\item{x}{an object}

\item{conv2dt}{What to convert to data.table}

\item{...}{
  Arguments passed on to \code{\link[=defactor_vector]{defactor_vector}}
  \describe{
    \item{\code{check.numeric}}{check if vector is numeric. Default is TRUE. Takes longer with this check but avoids type conversion (numeric to character).}
  }}
}
\value{
object of the same type without factors
}
\description{
Returns object of the same type without factors
}
