% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_defactor.r
\name{defactor_vector}
\alias{defactor_vector}
\title{Converts factor to character}
\usage{
defactor_vector(x, check.numeric = FALSE)
}
\arguments{
\item{x}{a vector}

\item{check.numeric}{check if vector is numeric. Default is TRUE. Takes longer with this check but avoids type conversion (numeric to character).}
}
\value{
character vector
}
\description{
Converts factor to character
}
