% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{detect_legal_form}
\alias{detect_legal_form}
\alias{magerman_detect_legal_form}
\title{Detect legal form}
\usage{
detect_legal_form(x, ...)
}
\arguments{
\item{x}{table}
}
\value{
standardized names table
}
\description{
Detect legal form
}
