% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{dots.and}
\alias{dots.and}
\title{Combines (with \code{&} funciton) two logical vectors. One is \code{arg.name} from dots the other is \code{arg.val}. Checks if they are the same length.}
\usage{
dots.and(
  arg.name,
  arg.val,
  env = parent.frame(),
  dots = eval(expression(list(...)), envir = env)
)
}
\arguments{
\item{arg.name}{Name of argument in dots to combine (&) \code{arg.val} with}

\item{arg.val}{Logical vector for combining with \code{arg.name} value in dots}

\item{env}{Do not set. This function needs its calling environment.}

\item{dots}{Do not set. This function needs ... (dots) values of calling}
}
\value{
Updated value for dots argument.
}
\description{
Combines (with \code{&} funciton) two logical vectors. One is \code{arg.name} from dots the other is \code{arg.val}. Checks if they are the same length.
}
