% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_patterns.r
\name{escape_regex}
\alias{escape_regex}
\title{Escapes special for regex characters}
\usage{
escape_regex(string)
}
\arguments{
\item{string}{character vector}
}
\value{
character vector with all special to regex characters escaped
}
\description{
Escapes special for regex characters
}
