% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_escape_patterns.r
\name{escape_regex_for_type}
\alias{escape_regex_for_type}
\title{Escapes special for different types of pattern}
\usage{
escape_regex_for_type(
  string,
  type,
  escape_fixed = FALSE,
  return_docs = FALSE,
  return_choices = FALSE,
  escape_begins = TRUE,
  escape_ends = TRUE
)
}
\arguments{
\item{string}{character vector (patterns)}

\item{type}{How the characters in patterns should be 'escaped' for
proper matching. The options are:}

\item{escape_fixed}{Whether to escape string for 'fixed' and
'exact' types and just trimmed string for "trim_exact". Default is FALSE}

\item{return_docs}{Return character vector of documentation for
each type (for using with roxygen2 documentations in many
places)}

\item{return_choices}{Return character vector of available escape
types}

\item{escape_begins}{Whether to escape string for 'begins' type}

\item{escape_ends}{Whether to escape string for 'ends' type}
}
\value{
character vector with all special to regex characters
escaped
}
\description{
Escapes special for different types of pattern

\itemize{
\item fixed  -  Match pattern string as it is within the target vector
\item begins  -  Match pattern string as it is in the beggining of the target vector
\item trim_begins  -  Match pattern string as it is in the beginning of the target vector ignoring preceding white-spaces
\item ends  -  Match pattern string as it is in the end of the target vector
\item trim_ends  -  Match pattern string as it is in the end of the target vector ignoring leading white-spaces
\item exact  -  Match pattern string exactly (i.e., match equal strings)
\item trim_exact  -  Match pattern string exactly (i.e., match equal strings) ignoring surrounding white-spaces
\item regex  -  Match regex pattern
}
}
