% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_patterns.r
\name{escape_regex_for_types}
\alias{escape_regex_for_types}
\title{Escapes special for regex characters conditionally}
\usage{
escape_regex_for_types(patterns, types, return_regex = FALSE)
}
\arguments{
\item{patterns}{character vector}

\item{types}{character vector of the same length as \code{strings} with
instructions whether and how to to escape regex characters. Options are:}

\item{return_regex}{Always return regex. Default is
true. Otherwise return unescaped string for 'fixed' and
'exact' types and just trimmed string for "trim_exact"}
}
\value{
string with all special to regex characters escaped
}
\description{
Escapes special for regex characters conditionally

\itemize{
\item fixed  -  Match pattern string as it is within the target vector
\item begins  -  Match pattern string as it is in the beggining of the target vector
\item trim_begins  -  Match pattern string as it is in the beginning of the target vector ignoring preceding white-spaces
\item ends  -  Match pattern string as it is in the end of the target vector
\item trim_ends  -  Match pattern string as it is in the end of the target vector ignoring leading white-spaces
\item exact  -  Match pattern string exactly (i.e., match equal strings)
\item trim_exact  -  Match pattern string exactly (i.e., match equal strings) ignoring surrounding white-spaces
\item regex  -  Match regex pattern
}
}
