% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_harmonize_options.r
\name{get_harmonize_options}
\alias{get_harmonize_options}
\title{Gets \code{harmonize_options} at point with consistent updates up through calling stack.}
\usage{
get_harmonize_options()
}
\value{
Returns list of updated arguments specified in \code{harmonize_options} function
}
\description{
Limited to max stack of 3 calls and calls that include at least \code{data} and \code{...} formals (\code{harmonizer} functions specific) up to \code{.GlobalEnv} or \code{harmonize} call.
}
