% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_get_standardize_options.r
\name{get_standardize_options}
\alias{get_standardize_options}
\title{Gets \code{standardize_options} at point with consistent updates up through calling stack.}
\usage{
get_standardize_options()
}
\value{
Returns list of updated arguments specified in \code{standardize_options} function
}
\description{
Limited to max stack of 3 calls and calls that include at least \code{data} and \code{...} formals (\code{nstandr} functions specific) up to \code{.GlobalEnv} or \code{standardize} call.
}
