% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.add.suffix}
\alias{harmonize.add.suffix}
\title{Adds a suffix to the string and counter at the end if needed}
\usage{
harmonize.add.suffix(
  name,
  suffix,
  x.names,
  search.suffix.in.name = TRUE,
  suffix.nbr.init = 1,
  suffix.nbr = NULL
)
}
\arguments{
\item{name}{Variable name}

\item{suffix}{Suffix}

\item{x.names}{Vector of variable names in x to check for duplicates and if we need to add a counter at the end}
}
\value{
Returns a new name
}
\description{
Adds a suffix to the string and counter at the end if needed
}
