% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.defactor}
\alias{harmonize.defactor}
\title{Defactor the object}
\usage{
harmonize.defactor(
  x,
  conv2dt = c("only.tables", "all.but.atomic", "all.but.lists", "all", "none"),
  ...
)
}
\arguments{
\item{x}{an object}

\item{conv2dt}{What to convert to data.table}

\item{...}{
  Arguments passed on to \code{\link[=harmonize.defactor.vector]{harmonize.defactor.vector}}
  \describe{
    \item{\code{check.numeric}}{check if vector is numeric. Default is TRUE. Takes longer with this check but avoids type conversion (numeric to character).}
  }}
}
\value{
object of the same type without factors
}
\description{
Returns object of the same type without factors
}
