% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.defactor.vector}
\alias{harmonize.defactor.vector}
\title{Converts factor to character}
\usage{
harmonize.defactor.vector(x, check.numeric = TRUE)
}
\arguments{
\item{x}{a vector}

\item{check.numeric}{check if vector is numeric. Default is TRUE. Takes longer with this check but avoids type conversion (numeric to character).}
}
\value{
character vector
}
\description{
Converts factor to character
}
