% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.detect}
\alias{harmonize.detect}
\title{This function is basically meant for coding names based on certain pattern}
\usage{
harmonize.detect(x, patterns, patterns.col = 1,
  patterns.type = "fixed", patterns.codes.col = 2, codes = NULL,
  codes.col = 1, codes.first = FALSE, codes.name = NA,
  codes.suffix = "coded", codes.omitted.val = NA,
  codes.append = TRUE, x.col = 1, x.codes.col = NULL,
  x.codes.update.empty = FALSE, x.codes.merge = FALSE,
  return.codes.only = FALSE, ...)
}
\arguments{
\item{x}{Vector or table to detect in.}

\item{patterns}{Accepts both vector or table. If patterns it is table can also include replacements column.}

\item{patterns.col}{If patterns is not a vector specifies which column to use. Default is 1.}

\item{patterns.type}{Kind of pattern. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling code{\link[stringi]{stri_replace_all_regex}})}

\item{patterns.codes.col}{If patterns is table which column to use as codes column.}

\item{codes}{If provided use it as codes. Should be the same length as patterns}

\item{codes.col}{If codes is not vector use this column for codes}

\item{codes.first}{If TRUE then return only codes for the first detected pattern. Otherwise return list of all matched codes. Default is FALSE.}

\item{codes.name}{If provided use it as a name for codes column in results.}

\item{codes.suffix}{If codes.name is not provided use this suffix to x.col name or x.vector.name if x is vector}

\item{x.codes.col}{If x is table, which column to use for making/merging/adding newly detected codes. Default is last column of x or NULL is x is vector}

\item{x.codes.update.empty}{If set then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).}

\item{x.codes.merge}{If set then merge (append) new codes to existing one.}

\item{return.codes.only}{If set then just return codes vector. Default is FALSE. Basically it resets return.x.cols to 0. So if it is set the return.x.cols (of harmonize.x helper) will be ignored.}

\item{...}{Arguments passed on to \code{harmonize.x}
\describe{
  \item{x}{an object}
  \item{x.inset}{a vector to inset. Optional. Default is NULL}
  \item{x.col}{vector of interest in \code{x} object}
  \item{x.rows}{Logical vector to filter records to harmonize. Default is NULL which means do not filter records}
  \item{x.rows.col}{Column that indicates which records to harmonize. If it is set then \code{x.rows} is ignored}
  \item{x.vector.name}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{x.col} will be used.}
  \item{x.harmonized.col}{Column in \code{x} where we want to put/update the \code{x.inset} vector. Default is NULL which means that we just put \code{x.inset} as a first vector/column and remove \code{x.col}. If set and \code{x.col} and \code{x.harmonized.col} are the same \code{x.col} wont be removed.}
  \item{x.harmonized.col.update}{Update values in this column if \code{x.rows} or \code{x.rows.col} is set. If set \code{harmonized.omitted.val} is ignored.}
  \item{harmonized.omitted.val}{If x.rows or x.rows.col is set. Use this value to fill the rest. Default is NA}
  \item{harmonized.append}{If set then put \code{x.inset} as the last instead of first vector/column. Default is FALSE.}
  \item{harmonized.name}{Use this name for the first column in results (harmonized names). Default is NULL, which means that either x.vector.name if x is vector or original x.col name will be used with \code{harmonized.suffix} at the end.}
  \item{harmonized.suffix}{If \code{harmonized.name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in \code{return.x.cols} it will add counter at the end to avoid variables with the same names.}
  \item{return.x.cols}{If x is table, set the columns to cbind to the result table. Default is -1, meaning cbind all but the first (original/unharmonized) column.}
  \item{return.x.cols.all}{Whether to bind all columns in x. Default is FALSE. If set the return.x.cols is ignored}
}}
}
\value{
If nothing was indicated to cbind to results then it returns harmonized vector. If something is needs to be cbind then it returns data.table
}
\description{
Optionally matches only at the beginning or at the end of the string.
}
