% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.escape.regex.cond}
\alias{harmonize.escape.regex.cond}
\title{Escapes special for regex characters conditionally}
\usage{
harmonize.escape.regex.cond(strings, conds)
}
\arguments{
\item{strings}{character vector}

\item{conds}{character vector of the same length as `strings` with instructions whether to escape regex ("fixed") add beginning ("begins") or ending ("ends") matcher. If value is "regex" then do not change the string}
}
\value{
string with all special to regex characters escaped
}
\description{
Escapes special for regex characters conditionally
}
