% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.escape.regex}
\alias{harmonize.escape.regex}
\title{Escapes special for regex characters}
\usage{
harmonize.escape.regex(string)
}
\arguments{
\item{string}{character vector}
}
\value{
character vector with all special to regex characters escaped
}
\description{
Escapes special for regex characters
}
