% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.escape.types}
\alias{harmonize.escape.types}
\title{Escapes special for regex characters conditionally}
\usage{
harmonize.escape.types(patterns, conds, all.regex = FALSE)
}
\arguments{
\item{conds}{character vector of the same length as \code{strings} with instructions whether to escape regex ("fixed") add beginning ("begins") or ending ("ends") matcher. Or if value is "regex" then do not change the string. Also possible to escape a regex for exact match ("exact") or exact match after trimming spaces ("trim.exact")}

\item{all.regex}{......}

\item{strings}{character vector}
}
\value{
string with all special to regex characters escaped
}
\description{
Escapes special for regex characters conditionally
}
