% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.is.empty}
\alias{harmonize.is.empty}
\title{Checks if elements that are either "", NA, NULL or have zero length}
\usage{
harmonize.is.empty(xs)
}
\arguments{
\item{xs}{vector}
}
\value{
logical vector of the same length
}
\description{
Checks if elements that are either "", NA, NULL or have zero length
}
