% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.is.ok.col}
\alias{harmonize.is.ok.col}
\title{Checks if a column(s) name/number is valid.}
\usage{
harmonize.is.ok.col(
  col,
  x,
  required = FALSE,
  allow.negative = FALSE,
  allow.zero = FALSE,
  several.ok = FALSE,
  ban.values = NULL,
  arg.name = deparse(substitute(col))
)
}
\arguments{
\item{col}{column name/number or vector of columns name/number}

\item{x}{table}

\item{required}{is set NULL is not allowed. Default is FALSE.}

\item{allow.negative}{If \code{col} is used for negation. Default is FALSE.}

\item{allow.zero}{Allow \code{col} to be 0. Default is FALSE.}

\item{several.ok}{If set \code{col} should refer to one column. Default is FALSE.}

\item{arg.name}{Name to use when reporting errors. By default set as \code{deparse(substitute(col))}}
}
\value{
TRUE if \code{col} value is ok and FALSE if it is NULL
}
\description{
Checks if a column(s) name/number is valid.
}
