% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.is.ok.dots}
\alias{harmonize.is.ok.dots}
\title{Checks if ... (dots) arguments are valid.}
\usage{
harmonize.is.ok.dots(dots.names, formals)
}
\arguments{
\item{dots.names}{Character vector of names of ... (dots) arguments. Usually obtained with \code{names(as.list(...))}.}

\item{formals}{Character vector of names to match dots agains. Usually obtained with \code{names(formals(function_name))}.}
}
\value{
TRUE if arguments are ok. FALSE if no arguments are provided (NULL or list())
}
\description{
Checks if ... (dots) arguments are valid.
}
