% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.is.ok.type}
\alias{harmonize.is.ok.type}
\title{Checks if object is valid type and length.}
\usage{
harmonize.is.ok.type(
  x,
  x.length = length(x),
  type = c("logical", "character", "numeric", "list", "atomic", NA),
  allow.na = TRUE,
  allow.null = TRUE,
  arg.name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Object to check.}

\item{x.length}{Length the object should adhere to. Default is objects length so it will always adhere.}

\item{type}{Type of the object. Default is "logical". If several types are provided that means that it cheches if the x is of either of types! (basically OR function)}

\item{allow.na}{Is NA allowed? Default is TRUE.}

\item{allow.null}{Is NULL allowed? Default is TRUE.}

\item{arg.name}{Name to use when reporting errors. By default set as \code{deparse(substitute(x))}}
}
\value{
TRUE if type is match, FALSE if x is NULL and it is allowed. Through an error otherwise
}
\description{
Checks if object is valid type and length.
}
