% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{harmonize.magerman}
\alias{harmonize.magerman}
\title{Harmonizes strings using exact procedures described in Magerman et al. 2009.}
\usage{
harmonize.magerman(
  x,
  magerman.procedures = magerman.procedures.table,
  detect.legal.form = FALSE,
  return.x.before.common.words.removal = FALSE,
  return.x.cols.all = FALSE,
  ...
)
}
\arguments{
\item{x}{table or vector}

\item{magerman.procedures}{list of procedures to pass to \code{harmonize} function. Default is \code{magerman.procedures.list}}

\item{detect.legal.form}{Whether to detect legal forms. Default is FALSE}

\item{return.x.before.common.words.removal}{Whether to save harmonized column before \code{common.words.removal} procedure. Default is FALSE}

\item{return.x.cols.all}{Whether to return initial column in x. Default is FALSE}

\item{...}{
  Arguments passed on to \code{\link[=harmonize]{harmonize}}
  \describe{
    \item{\code{procedures}}{List of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments.}
    \item{\code{progress}}{Show the progress? Default is TRUE}
    \item{\code{progress.min}}{The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5}
    \item{\code{progress.by}}{If set it will divide the x into chunk of this amount of rows. Default is NA.}
    \item{\code{progress.percent}}{Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.}
    \item{\code{progress.message.use.names}}{Should we use names from \code{procedures} list to report progress. Default is TRUE.}
    \item{\code{quite}}{Suppress all messages. Default is FALSE.}
  }}
}
\value{
Harmonized table
}
\description{
Harmonizes strings using exact procedures described in Magerman et al. 2009.
}
\references{
Magerman et al., 2006 - Data Production Methods for Harmonized Patent Statistics: Patentee Name Harmonization
}
