% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.make.procedures.list}
\alias{harmonize.make.procedures.list}
\title{Makes list of procedures calls from table.}
\usage{
harmonize.make.procedures.list(
  procedures.table,
  message.field = "message",
  function.call.field = "function.call",
  no.field = "no",
  remove.comments = TRUE,
  sort.by.no.field = TRUE,
  comments = c("#", "-", "")
)
}
\arguments{
\item{procedures.table}{Table to use}

\item{message.field}{name of the column with messages that will be displayed when each call is executed}

\item{function.call.field}{name of the column where function (harmonization procedures) calls are listed.}

\item{no.field}{name of the column where the number of procedure is specified. Also this field indicates if the row in the table is just a comment in which case it will be removed if \code{remove.comments} is set (which is set by default)}

\item{remove.comments}{Whether to remove comments.}
}
\value{
List of named function calls. Names are messages.
}
\description{
Table should have at least two columns - messages and fuctions calls. Each function call should be a string of the following format "'function.name', arg1 = val1, arg2 = val2" (same as arguments for \code{do.call} function).
}
