% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.match.arg}
\alias{harmonize.match.arg}
\title{Matches the argument vector to (default) choices and ensures the correct length}
\usage{
harmonize.match.arg(
  arg,
  arg.length = 1,
  arg.length.check = TRUE,
  ensure.length = TRUE,
  choices = NULL,
  arg.call = substitute(arg),
  env = parent.frame()
)
}
\arguments{
\item{arg}{An argument vector to check if it is matches the values}

\item{arg.length}{Desired length of the \code{arg} to check against or to ensure}

\item{arg.length.check}{Whether to check \code{arg} length}

\item{ensure.length}{Whether to repeat \code{arg} \code{length} times if \code{arg} is length of 1}

\item{choices}{Vector of values to match \code{arg}}

\item{arg.call}{Saves \code{arg} call. Do not touch!}

\item{env}{Saves environment where the function was called from. Do not touch!}

\item{length}{Check if the \code{arg} is this lenght. If \code{arg} is length of 1 and \code{ensure.length} is set to TRUE (default) then it will repeat \code{arg} \code{length} times and return}
}
\value{
Argument vector
}
\description{
Matches the argument vector to (default) choices and ensures the correct length
}
