% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.unlist.column}
\alias{harmonize.unlist.column}
\title{If column in the `x` table is list unlist it if possible}
\usage{
harmonize.unlist.column(x)
}
\arguments{
\item{x}{object}
}
\value{
updated object
}
\description{
If column in the `x` table is list unlist it if possible
}
