% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.x.length}
\alias{harmonize.x.length}
\title{Gets lengths of the object}
\usage{
harmonize.x.length(x)
}
\arguments{
\item{x}{object (table)}
}
\value{
Width (nrow) of the object. If it is atomic it returns its length.
}
\description{
Gets lengths of the object
}
