% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.x.split}
\alias{harmonize.x.split}
\title{Splits the object (table) in chunks by rows}
\usage{
harmonize.x.split(x, by, len)
}
\arguments{
\item{x}{object or table}

\item{by}{number of rows to split by}

\item{len}{length of the table (nrow)}
}
\value{
List of (sub)tables
}
\description{
Convenient to apply some function to the table in chunks, e.g., if you want to add display of progress.
}
