% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.x.width}
\alias{harmonize.x.width}
\title{Gets width of the object}
\usage{
harmonize.x.width(x)
}
\arguments{
\item{x}{object (table)}
}
\value{
Width (ncol) of the object. If it is atomic it is 1.
}
\description{
Gets width of the object
}
