% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cockburn.r
\name{harmonize_cockburn}
\alias{harmonize_cockburn}
\title{Harmonizes strings using exact procedures described in Cockburn, et al. (2009)}
\usage{
harmonize_cockburn(
  x,
  cockburn_procedures = cockburn_procedures_table,
  detect_legal_form = FALSE,
  return_x_before_common_words_removal = FALSE,
  ...
)
}
\arguments{
\item{x}{table or vector}

\item{cockburn_procedures}{list of procedures to pass to \code{harmonize} function. Default is \code{cockburn_procedures.list}}

\item{detect_legal_form}{Whether to detect legal forms. Default is FALSE}

\item{return_x_before_common_words_removal}{Whether to save harmonized column before \code{common.words.removal} procedure. Default is FALSE}

\item{...}{
  Arguments passed on to \code{\link[=harmonize]{harmonize}}
  \describe{
    \item{\code{procedures}}{Named list of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments. Names of the list elements are used for progress messages. Procedures can also be passed as data.frame in which case it will be converted to list of procedures with \code{harmonize_make_procedures_list} (see its help for the correct format of data.frame with procedures). Default is \code{harmonizer_default_procedures_table}}
    \item{\code{show_progress}}{Whether to report progress percentage. Default is TRUE}
    \item{\code{nrows_min_to_show_progress}}{The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5}
    \item{\code{progress_step_nrows}}{If set it will divide the x into chunk of this amount of rows. Default is NULL.}
    \item{\code{progress_step_in_percent}}{Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.}
    \item{\code{progress_message_use_names}}{Should we use names from \code{procedures} list to report progress. Default is TRUE.}
    \item{\code{quite}}{Suppress all messages. Default is FALSE.}
    \item{\code{save_intermediate_x_to_var}}{For debuging of standartization procedures. Saves intermediate results to this variable. If procedures finish without errors then the variable will be removed.}
  }}
}
\value{
Harmonized table
}
\description{
Harmonizes strings using exact procedures described in Cockburn, et al. (2009)
}
\references{
Cockburn, et al. (2009)
}
