% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_data_length}
\alias{harmonize_data_length}
\title{Gets lengths of the object}
\usage{
harmonize_data_length(data)
}
\arguments{
\item{data}{input data (table)}
}
\value{
Length (\code{nrow}) of the object. If it is atomic it returns its length.
}
\description{
Gets lengths of the object
}
