% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_data_width}
\alias{harmonize_data_width}
\title{Gets width of the object}
\usage{
harmonize_data_width(data)
}
\arguments{
\item{data}{object (table)}
}
\value{
Width (ncol) of the object. If it is atomic it is 1.
}
\description{
Gets width of the object
}
