% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_dehtmlize}
\alias{harmonize_dehtmlize}
\title{Converts HTML characters to UTF-8}
\usage{
harmonize_dehtmlize(
  x,
  as_single_string = FALSE,
  as_single_string_sep = "#_|",
  use_read_xml = FALSE,
  ...
)
}
\arguments{
\item{x}{object (table)}

\item{as_single_string}{If set then collapse characters in the main column of the \code{x} (i.e., \code{x.col}) as to a single string. It will increase performance (at least for relatively short tables). Default is FALSE}

\item{as_single_string_sep}{delimiter for collapsed strings to uncollapse it later. Default is "#_|".}

\item{use_read_xml}{If set the it will parse XML. Default is FALSE which means it parses HTML}

\item{...}{
  Arguments passed on to \code{\link[=harmonize_options]{harmonize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to harmonize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (harmonized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{x_atomic_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_col_name}}{Use this name for the column with results (harmonized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}
}
\value{
updated object
}
\description{
The method is about 1/3 faster than htmlParse but it is still quite slow
}
\references{
http://stackoverflow.com/questions/5060076
}
