% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_detect.r
\name{harmonize_detect}
\alias{harmonize_detect}
\title{This function is basically meant for coding names based on certain pattern}
\usage{
harmonize_detect(
  x,
  patterns,
  patterns.col = 1,
  patterns.as.codes = FALSE,
  patterns.codes.col = 2,
  patterns.type = c("fixed", "begins", "begins.trimmed", "ends", "ends.trimmed",
    "regex", "exact", "exact.trimmed"),
  patterns.type.col = NULL,
  codes = NULL,
  codes.name = NA,
  codes.suffix = "coded",
  codes.omitted.val = NA,
  codes.prepend = FALSE,
  x.codes.col = NULL,
  x.codes.update.empty = FALSE,
  x.codes.merge = FALSE,
  x.codes.merge.prepend = FALSE,
  return.only.codes = FALSE,
  return.only.first.detected.code = FALSE,
  ...
)
}
\arguments{
\item{x}{Vector or table to detect in.}

\item{patterns}{Accepts both vector or table. If patterns it is table can also include replacements column.}

\item{patterns.col}{If patterns is not a vector specifies which column to use. Default is 1.}

\item{patterns.as.codes}{If set then use patterns as codes. In that case codes are ignored and there is no need to provide them. Default is FALSE.}

\item{patterns.codes.col}{If patterns is table which column to use as codes column. Default is 2.}

\item{patterns.type}{Specifies kind(s) of patterns. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling code{\link[stringi]{stri_replace_all_regex}}). Another possible options are "exact" and "trim.exact" which matches a pattern string exactly or its trimmed version (i.e., white spaces are ignored at the beginning and at the end of the pattern)}

\item{patterns.type.col}{Column in patterns table where you can specify types of patterns for each pattern. If set then \code{patterns.type} is ignored. Default is NULL}

\item{codes}{If provided use it as codes. Should be the same length as patterns. Default is NULL.}

\item{codes.name}{If provided use it as a name for codes column in results. It also will work as \code{x.codes.col} if \code{x.codes.merge} or \code{x.codes.update.empyty} are set. In case if \code{x.codes.merge} or \code{x.codes.update.empyty} are set and names(\code{x}) does not have column name \code{codes.name} it will reset both \code{x.codes.merge} and \code{x.codes.update.empyty} to FALSE and will (pre)append a new column \code{codes.name}.}

\item{codes.suffix}{If \code{codes.name} is not provided use this suffix to append to \code{x.col} name or \code{x.atomic.name} if x is vector (see harmonize.x function).}

\item{codes.omitted.val}{If \code{x.rows} is set. Use this value to fill the omitted rows. Default is vector of NAs of \code{x} length. When we update existing codes column (i.e., \code{x.codes.col} is set and valid) and the value is not set explicitly it resets to NULL which means that we use initial values for omitted rows.}

\item{codes.prepend}{Whether to prepend codes column to the \code{x}. Default is FALSE. If set it will prepend codes column to \code{x} instead of append.}

\item{x.codes.col}{If \code{x} is table, which column to use for making/merging/adding newly detected codes. Default is NULL. When it is not set explicitly and \code{x.codes.merge} or \code{x.codes.update.empty} are set it will first try to use \code{codes.name} if it is provided or if not will set the last column of \code{x} as \code{x.codes.col}.}

\item{x.codes.update.empty}{If set then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).}

\item{x.codes.merge}{If set then merge (append or prepend) new codes to existing ones. It might turn the codes column to list of vectors.}

\item{x.codes.merge.prepend}{If set then it will prepend new codes if not then it will append (default). It is ignored if \code{x.codes.merge} is not set.}

\item{return.only.codes}{If set then just return codes vector. Default is FALSE. Basically when it is set it changes \code{return.x.cols} to 0.}

\item{return.only.first.detected.code}{If TRUE then return only codes for the first detected pattern. If FALSE return list of vectors of all matched codes. Default is TRUE.}

\item{...}{
  Arguments passed on to \code{\link[=harmonize.x]{harmonize.x}}
  \describe{
    \item{\code{inset.vector}}{a vector to inset. Optional. Default is NULL}
    \item{\code{x.col}}{vector of interest in \code{x} object}
    \item{\code{x.col.update}}{Update values in \code{x.col} column. Default is FALSE. If set \code{inset.append}, \code{inset.name} and \code{inset.suffix} are ignored. Also if set the default for \code{return.x.cols.all} will be set to TRUE.}
    \item{\code{x.rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records}
    \item{\code{x.atomic.name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{x.col} will be used.}
    \item{\code{inset.omitted.val}}{If \code{x.rows} is set merge these values to the results. It should be a vector of length 1 or \code{nrow(x)}. If the value is NULL (default) then use values of \code{x.col}.}
    \item{\code{inset.append}}{If set then put \code{inset.vector} as the last instead of first vector/column. Default is FALSE.}
    \item{\code{inset.suffix}}{If \code{inset.name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in \code{return.x.cols} it will add counter at the end to avoid variables with the same names.}
    \item{\code{inset.name}}{Use this name for the first column in results (harmonized names). Default is NULL, which means that either x.atomic.name if x is vector or original x.col name will be used with \code{inset.suffix} at the end.}
    \item{\code{return.x.cols}}{If x is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (x.col).}
    \item{\code{return.x.cols.all}}{Whether to bind all columns in x. Defaults depends on values of \code{x.col.update} and \code{inset.append}. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the return.x.cols is ignored.}
  }}
}
\value{
The updated \code{x} table with codes column or just codes if \code{return.only.codes} is set.
}
\description{
Optionally matches only at the beginning or at the end of the string.
}
