% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_empty_omit}
\alias{harmonize_empty_omit}
\title{Removes elements that are either "", NA, NULL or have zero length}
\usage{
harmonize_empty_omit(x)
}
\arguments{
\item{x}{vector}
}
\value{
updated vector with empty elements removed
}
\description{
Removes elements that are either "", NA, NULL or have zero length
}
