% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_escape_regex}
\alias{harmonize_escape_regex}
\title{Escapes special for regex characters}
\usage{
harmonize_escape_regex(string)
}
\arguments{
\item{string}{character vector}
}
\value{
character vector with all special to regex characters escaped
}
\description{
Escapes special for regex characters
}
