% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_escape_type}
\alias{harmonize_escape_type}
\title{Escapes special for different types of pattern}
\usage{
harmonize_escape_type(
  string,
  type = c("fixed", "begins", "begins.trimmed", "ends", "ends.trimmed", "regex",
    "exact", "exact.trimmed"),
  all.regex = TRUE
)
}
\arguments{
\item{string}{character vector}

\item{type}{whether it should escape regex ("fixed") add beginning
("begins") or ending ("ends") matcher. Or if value is "regex"
then do not change the string. Also possible to escape a regex
for exact match ("exact") or exact match after trimming spaces
("trim.exact")}
}
\value{
character vector with all special to regex characters
escaped
}
\description{
Escapes special for different types of pattern
}
