% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_is_data_empty}
\alias{harmonize_is_data_empty}
\title{Checks if all elements in vercor(s) are either "", NA, NULL or have zero length}
\usage{
harmonize_is_data_empty(data)
}
\arguments{
\item{data}{input data to check each vector}
}
\value{
logical vector of the same length
}
\description{
Checks if all elements in vercor(s) are either "", NA, NULL or have zero length
}
