% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_utils.r
\name{harmonize_is_ok_dots}
\alias{harmonize_is_ok_dots}
\title{Checks if ... (dots) arguments are valid.}
\usage{
harmonize_is_ok_dots(dots.names, formals)
}
\arguments{
\item{dots.names}{Character vector of names of ... (dots) arguments. Usually obtained with \code{names(as.list(...))}.}

\item{formals}{Character vector of names to match dots agains. Usually obtained with \code{names(formals(function_name))}.}
}
\value{
TRUE if arguments are ok. FALSE if no arguments are provided (NULL or list())
}
\description{
Checks if ... (dots) arguments are valid.
}
