% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_options}
\alias{harmonize_options}
\title{This function provides options for \code{harmonizer} functions that controls how the input \code{data} object, intermediate and end results  are handled.}
\usage{
harmonize_options(
  col = 1,
  rows = NULL,
  ommitted_rows_values = NULL,
  placement = c("replace_col", "prepend_to_data", "append_to_data", "prepend_to_col",
    "append_to_col", "ignore"),
  name = NA,
  name_if_input_atomic = "names",
  name_suffix = "_harmonized",
  append_copy = FALSE,
  append_copy_name_format = "\%name_harmonizing_\%number_\%procedure",
  select_cols = NA
)
}
\arguments{
\item{col}{Column of interest in the \code{data} object. The one we need to harmonize}

\item{rows}{Logical vector to filter records of interest. Default is NULL which means do not filter records}

\item{ommitted_rows_values}{If \code{rows} is set merge these values to the results. It should be a vector of length 1 or \verb{nrow(data). If the value is NULL (default) then use original values of }col`}

\item{placement}{Where to inset retults (harmonized vector) in the \code{data} object. Default options is "replace_col" which overwrides the \code{col} in \code{data} with results.}

\item{name}{Use this name for the first column in results (harmonized names). Default is NULL, which means that either name_if_data_atomic if data is vector or original col name will be used with \code{name_suffix} at the end.}

\item{name_if_input_atomic}{If \code{data} is vector use this name for original column if it is in results. Default is "data". If \code{data} is table the name of \code{col} will be used.}

\item{name_suffix}{If \code{name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already edataists in \code{select_data_cols} it will add counter at the end to avoid variables with the same names.}

\item{append_copy}{Whether to append a copy of result vector to \code{data} object}

\item{append_copy_name_format}{How the append copy wiil be named}

\item{select_cols}{If data` object is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (col).}
}
\value{
list of options updated with arguments from \code{harmonizer} functions calls stack up to .GlobalEnv or 'harmonize' call.
}
\description{
It solves problem of passing options parameters to some arbitrary harmonizing function that are also passed with \code{...} (in which case there will be an error that parameters have duplicated names). Using this function allows not to pass \code{...} explicitly down to calling g stack and get the options values updated by explicit parameters at any location in the stack.
}
