% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_squish_spaces.r
\name{harmonize_squish_spaces}
\alias{harmonize_squish_spaces}
\title{Removes redundant whitespases}
\usage{
harmonize_squish_spaces(x, wrap_in_spaces = FALSE, ...)
}
\arguments{
\item{x}{table or vector}

\item{wrap_in_spaces}{If set then adds leaing and ending spaces. Default is FALSE.}

\item{...}{
  Arguments passed on to \code{\link[=harmonize_options]{harmonize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to harmonize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (harmonized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{x_atomic_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_col_name}}{Use this name for the column with results (harmonized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}
}
\value{
updated table or vector
}
\description{
Removes redundant whitespases
}
