% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize_target_get}
\alias{harmonize_target_get}
\title{Gets a target vector to harmonize.}
\usage{
harmonize_target_get(data, target_col, target_rows)
}
\arguments{
\item{data}{Input data. Can be vector, data.frame or a data.table}

\item{target_col}{Column of interest in the input \code{data}. The vector we would like to work with. This parameter is ignored if input \code{data} is a vector (checked by \code{is.atomic})}
}
\value{
A vector. Factors in imput \code{data} are converted to string.
}
\description{
Gets a target vector to harmonize.
}
