% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/harmonizer.r
\name{harmonizer}
\alias{harmonizer}
\title{Harmonizes organizational names.}
\usage{
harmonizer(org.names, procedures = list("toutf", "tolower", "html2txt",
  "clear.in.brackets", "toascii", "toupper", "standardize.nber", "trims"))
}
\arguments{
\item{org.names}{Character vector of organizational names to harmonize}

\item{procedures}{List of harmonization procedures. Each procedure can be specified as a string representing procedure name (see details for procedure names) or as a list where the first element should be procedure name (string) and other elements will passed as arguments to this procedure.}
}
\value{
Character vector of harmonized  organizational names
}
\description{
Returns harmonized version of organizational names.
}
\details{
The following procedures are available:
\itemize{
\item toutf - encode as UTF-8 (a wrapper for str_conv(str, "UTF-8"))
\item tolower - lowercase
\item toupper - uppercase
\item html2txt - remove HTML symbols and tags (takes lot of time)
\item toascii - replace accented characters with ascii equivalent
\item clear.in.brackets - removes brackets and its content - (), [], {}, <>
\item trims - removes double white spaces and trims white spaces
\item standardize.nber - applies standard name substitutions following NBER's PDP (Thoma, et al, 2010) - https://sites.google.com/site/patentdataproject/Home/posts/namestandardizationroutinesuploaded
\item standardize.nber.sansremovals - same as standardize.nber but without removals of organization legal form
\item standardize.nber.sansderwent - same as standardize.nber but without Derwent names standardization
\item standardize.nber.sansderwent.sansremovals - same as standardize.nber.sansderwent but without removals of organization legal form
\item standardize.punctuation - removes punctuation
}
}
\examples{
org.names.test <- c("žŸong-ÂÃÇÈÏa\\n\\u00b5 &oacute;\\u00b5<p>,  INt LTD &AMP; Co;  "
                  , "<br> the $ (&AMP; C&oacute;MP comPANY) Ïotta")
data.frame(original = org.names.test
         , harmonized = harmonize(org.names.test))
}
