% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{inset_vector}
\alias{inset_vector}
\title{Insets target vector back to input object (\code{data})}
\usage{
inset_vector(
  data,
  vector = NULL,
  col = 1,
  update = FALSE,
  append = FALSE,
  rows = NULL,
  ommitted_rows_values = NULL,
  name = NA,
  name_suffix = "harmonized",
  name_if_data_atomic = "names",
  select_data_cols = -ifelse(is.numeric(col), col, match(col, names(data))),
  select_all_data_cols = append | update
)
}
\arguments{
\item{data}{an object}

\item{vector}{a vector to inset. Optional. Default is NULL}

\item{col}{vector of interest in \code{data} object}

\item{update}{Update values in \code{col} column. Default is FALSE. If set \code{append}, \code{name} and \code{name_suffix} are ignored. Also if set the default for \code{select_all_data_cols} will be set to TRUE.}

\item{append}{If set then put \code{vector} as the last instead of first vector/column. Default is FALSE.}

\item{rows}{Logical vector to filter records of interest. Default is NULL which means do not filter records}

\item{ommitted_rows_values}{If \code{rows} is set merge these values to the results. It should be a vector of length 1 or \code{nrow(data)}. If the value is NULL (default) then use original values of \code{col}.}

\item{name}{Use this name for the first column in results (harmonized names). Default is NULL, which means that either name_if_data_atomic if data is vector or original col name will be used with \code{name_suffix} at the end.}

\item{name_suffix}{If \code{name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already edataists in \code{select_data_cols} it will add counter at the end to avoid variables with the same names.}

\item{name_if_data_atomic}{If \code{data} is vector use this name for original column if it is in results. Default is "data". If \code{data} is table the name of \code{col} will be used.}

\item{select_data_cols}{If data is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (col).}

\item{select_all_data_cols}{Whether to bind all columns in \code{data}. Defaults depends on values of \code{update} and \code{append}. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the select_data_cols is ignored.}
}
\value{
returns updated \code{data} object
}
\description{
Insets target vector back to input object (\code{data})
}
