% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{inset_vector_new}
\alias{inset_vector_new}
\title{Insets target vector back to input object (\code{data})}
\usage{
inset_vector_new(data, vector, ...)
}
\arguments{
\item{data}{an object}

\item{vector}{a vector to inset. Optional. Default is NULL}

\item{...}{
  Arguments passed on to \code{\link[=harmonize_options]{harmonize_options}}
  \describe{
    \item{\code{col}}{Column of interest in the \code{data} object. The one we need to harmonize}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records}
    \item{\code{ommitted_rows_values}}{If \code{rows} is set merge these values to the results. It should be a vector of length 1 or \verb{nrow(data). If the value is NULL (default) then use original values of }col`}
    \item{\code{placement}}{Where to inset retults (harmonized vector) in the \code{data} object. Default options is "replace_col" which overwrides the \code{col} in \code{data} with results.}
    \item{\code{name}}{Use this name for the first column in results (harmonized names). Default is NULL, which means that either name_if_data_atomic if data is vector or original col name will be used with \code{name_suffix} at the end.}
    \item{\code{name_if_input_atomic}}{If \code{data} is vector use this name for original column if it is in results. Default is "data". If \code{data} is table the name of \code{col} will be used.}
    \item{\code{name_suffix}}{If \code{name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already edataists in \code{select_data_cols} it will add counter at the end to avoid variables with the same names.}
    \item{\code{append_copy}}{Whether to append a copy of result vector to \code{data} object}
    \item{\code{append_copy_name_format}}{How the append copy wiil be named}
    \item{\code{select_cols}}{If data` object is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (col).}
  }}
}
\value{
returns updated \code{data} object
}
\description{
Insets target vector back to input object (\code{data})
}
