% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{is_empty}
\alias{is_empty}
\title{Checks if string has something to print}
\usage{
is_empty(x)
}
\arguments{
\item{x}{string to test}
}
\value{
boolean
}
\description{
Checks if string has something to print
}
