% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman.detect.comma.period.irregularities}
\alias{magerman.detect.comma.period.irregularities}
\title{Detects comma period irregularities}
\usage{
magerman.detect.comma.period.irregularities(x, ...)
}
\arguments{
\item{x}{table}

\item{...}{Arguments passed on to \code{harmonize.detect}
\describe{
  \item{x}{Vector or table to detect in.}
  \item{patterns}{Accepts both vector or table. If patterns it is table can also include replacements column.}
  \item{patterns.col}{If patterns is not a vector specifies which column to use. Default is 1.}
  \item{patterns.type}{Kind of pattern. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling code{\link[stringi]{stri_replace_all_regex}})}
  \item{patterns.codes.col}{If patterns is table which column to use as codes column.}
  \item{codes}{If provided use it as codes. Should be the same length as patterns}
  \item{codes.col}{If codes is not vector use this column for codes}
  \item{codes.name}{If provided use it as a name for codes column in results.}
  \item{codes.suffix}{If codes.name is not provided use this suffix to x.col name or x.vector.name if x is vector}
  \item{codes.first}{If TRUE then return only codes for the first detected pattern. Otherwise return list of all matched codes. Default is FALSE.}
  \item{x.codes.col}{If x is table, which column to use for making/merging/adding newly detected codes. Default is last column of x or NULL is x is vector}
  \item{x.codes.update.empty}{If set then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).}
  \item{x.codes.merge}{If set then merge (append) new codes to existing one.}
  \item{return.codes.only}{If set then just return codes vector. Default is FALSE. Basically it resets return.x.cols to 0. So if it is set the return.x.cols (of harmonize.x helper) will be ignored.}
}}
}
\value{
Harmonized table
}
\description{
Detects comma period irregularities
}
