% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman.remove.legal.form.and.clean}
\alias{magerman.remove.legal.form.and.clean}
\title{Removes legal form}
\usage{
magerman.remove.legal.form.and.clean(x)
}
\arguments{
\item{x}{table}
}
\value{
Harmonized table
}
\description{
Removes legal form
}
