% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman.remove.legal.form}
\alias{magerman.remove.legal.form}
\title{Removes legal form}
\usage{
magerman.remove.legal.form(x)
}
\arguments{
\item{x}{table}
}
\value{
Harmonized table
}
\description{
Removes legal form
}
