% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman.replace.proprietary.characters}
\alias{magerman.replace.proprietary.characters}
\title{Replaces proprietary characters}
\usage{
magerman.replace.proprietary.characters(x, ...)
}
\arguments{
\item{x}{table}

\item{...}{Arguments passed on to \code{harmonize.replace}
\describe{
  \item{x}{Vector or table to harmonize.}
  \item{patterns}{Accepts both vector or table. If patterns it is table can also include replacements column.}
  \item{patterns.col}{If patterns is not a vector which column to use. Default is 1.}
  \item{patterns.type}{Kind of pattern. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling code{\link[stringi]{stri_replace_all_regex}})}
  \item{patterns.type.col}{}
  \item{patterns.replacements.col}{If patterns is not a vector and includes replacements which column to use for replacements. Default is 2.}
  \item{replacements}{If patterns does not have column with replacements provide it here.}
  \item{replacements.col}{If replacements is not a vector which column to use. Default is 1.}
}}
}
\value{
Harmonized table
}
\description{
Replaces proprietary characters
}
