% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman_detect_characters}
\alias{magerman_detect_characters}
\title{Detect candidates for characters that need to be cleaned}
\usage{
magerman_detect_characters(
  x,
  output_codes_col_name = "characters_cleaning_candidates",
  ...
)
}
\arguments{
\item{x}{table}

\item{output_codes_col_name}{Same as in \code{detect_patterns} (see below)}

\item{...}{
  Arguments passed on to \code{\link[=detect_patterns]{detect_patterns}}
  \describe{
    \item{\code{patterns}}{Accepts both vector or table. If patterns it is table can also include replacements column.}
    \item{\code{patterns_col}}{If patterns is a table this specifies which column to use. Default is 1.}
    \item{\code{patterns_codes_col}}{If patterns is table which column to use as codes column. Default is 2.}
    \item{\code{patterns_type}}{Specifies kind(s) of patterns. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are:}
    \item{\code{patterns_type_col}}{Column in patterns table where you can specify types of patterns for each pattern. If set then \code{patterns.type} is ignored. Default is NULL.}
    \item{\code{patterns_codes}}{If provided use it as codes. Should be the same length as patterns. Default is NULL.}
    \item{\code{codes_omitted_rows_value}}{If \code{rows} is set. Use this value to fill the omitted rows. When we update existing codes column Default is NULL which means that we use initial codes values for omitted rows. If there is no codes col to update omited rows will be filled with NA.}
    \item{\code{no_match_code}}{If provided code records that did not get any match with it.}
    \item{\code{merge_existing_codes}}{Whether to merge newly detected codes with existing. Options are:}
    \item{\code{return_only_codes}}{If toggled on then just return codes vector.}
    \item{\code{return_only_first_detected_code}}{If TRUE then return only codes for the first detected pattern. If FALSE return list of all matched codes. Default is TRUE. (Currently does affect performance)}
    \item{\code{return_merge_codes_description}}{Return description of choices for \code{merge_existing_codes} paramenter.}
  }}
}
\value{
coded table
}
\description{
Detect candidates for characters that need to be cleaned
}
