% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman_detect_legal_form}
\alias{magerman_detect_legal_form}
\title{Detects legal form}
\usage{
magerman_detect_legal_form(x, ...)
}
\arguments{
\item{x}{table}

\item{...}{
  Arguments passed on to \code{\link[=magerman_detect_legal_form_end]{magerman_detect_legal_form_end}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Harmonized table
}
\description{
Detects legal form
}
