% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman_detect_umlaut}
\alias{magerman_detect_umlaut}
\title{Detects umlauts}
\usage{
magerman_detect_umlaut(x, ...)
}
\arguments{
\item{x}{table}

\item{...}{
  Arguments passed on to \code{\link[=detect_patterns]{detect_patterns}}
  \describe{
    \item{\code{patterns}}{Accepts both vector or table. If patterns it is table can also include replacements column.}
    \item{\code{patterns_col}}{If patterns is a table this specifies which column to use. Default is 1.}
    \item{\code{patterns_codes_col}}{If patterns is table which column to use as codes column. Default is 2.}
    \item{\code{patterns_type}}{Specifies kind(s) of patterns. Default is "fixed" (calling code{\link[stringi]{stri_replace_all_fixed}}). Other options are:}
    \item{\code{patterns_type_col}}{Column in patterns table where you can specify types of patterns for each pattern. If set then \code{patterns.type} is ignored. Default is NULL.}
    \item{\code{patterns_codes}}{If provided use it as codes. Should be the same length as patterns. Default is NULL.}
    \item{\code{codes_col_name}}{If provided use it as a name for codes column (new if it does not exist or the one to update) in results.}
    \item{\code{codes_col_name_suffix}}{If \code{codes_col_name} is not provided use this suffix to append to \code{col} name or \code{x_atomic_name}  (see\code{harmonize_options}).}
    \item{\code{codes_omitted_rows_value}}{If \code{rows} is set. Use this value to fill the omitted rows. When we update existing codes column Default is NULL which means that we use initial codes values for omitted rows. If there is no codes col to update omited rows will be filled with NA.}
    \item{\code{no_match_code}}{If provided code records that did not get any match with it.}
    \item{\code{codes_update_empty}}{If toggled on then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).}
    \item{\code{codes_merge}}{If set then merge (append or prepend) new codes to existing ones. It might turn the codes column to list of vectors.}
    \item{\code{codes_merge_prepend}}{If set then it will prepend new codes if not then it will append (default). It is ignored if \code{x.codes.merge} is not set.}
    \item{\code{return_only_codes}}{If toggled on then just return codes vector.}
    \item{\code{return_only_first_detected_code}}{If TRUE then return only codes for the first detected pattern. If FALSE return list of all matched codes. Default is TRUE. (Currently does affect performance)}
  }}
}
\value{
Coded table
}
\description{
Detects umlauts
}
