% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman_remove_legal_form_and_clean}
\alias{magerman_remove_legal_form_and_clean}
\title{Removes legal form}
\usage{
magerman_remove_legal_form_and_clean(x)
}
\arguments{
\item{x}{table}
}
\value{
Harmonized table
}
\description{
Removes legal form
}
