% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman_replace_umlaut}
\alias{magerman_replace_umlaut}
\title{Replaces Umlauts}
\usage{
magerman_replace_umlaut(
  x,
  has_umlaut_col = NULL,
  drop_has_umlaut_col = TRUE,
  replace_accented_characters = FALSE,
  ...
)
}
\arguments{
\item{x}{Data}

\item{has_umlaut_col}{Column with logical values indicating weather a corresponding string has an umlaut. Default is NULL so it detects is automatically first}

\item{drop_has_umlaut_col}{Whether to drop \code{has_umlaut_col}. Default is FALSE}

\item{replace_accented_characters}{Whether to replace accented characters first. Default is FALSE}

\item{...}{
  Arguments passed on to \code{\link[=standardize_options]{standardize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to standardize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (standardized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{x_atomic_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_col_name}}{Use this name for the column with results (standardized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}
}
\value{
standardized names table
}
\description{
A simple illustration of what this procedure does:

\if{html}{\out{<div class="sourceCode">}}\preformatted{GLÜHLAMPEN -> GLUEHLAMPEN
}\if{html}{\out{</div>}}
}
