% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{make_dot_edges}
\alias{make_dot_edges}
\title{Makes dot graph edges for visualizing arrows between sequence of procedures.}
\usage{
make_dot_edges(procedures_list, prefix = NULL)
}
\arguments{
\item{procedures_list}{List of procedures. See \code{sandardize} for how to specify of procedures list.}

\item{prefix}{Prefix for edge IDs}
}
\value{
list of dot graph edges
}
\description{
Makes dot graph edges for visualizing arrows between sequence of procedures.
}
