% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{make_dot_graph}
\alias{make_dot_graph}
\title{Generates graph description for visualizing list of procedures in dot format.}
\usage{
make_dot_graph(procedures_list, ...)
}
\arguments{
\item{procedures_list}{List of procedures. See \code{sandardize} for how to specify of procedures list.}

\item{...}{
  Arguments passed on to \code{\link[=make_dot_nodes]{make_dot_nodes}}
  \describe{
    \item{\code{node_prefix}}{pefix for node IDs}
    \item{\code{procedures_group_bg_color}}{Background color for visuallizing groupped procedures (nested list of procedures)}
    \item{\code{procedures_group_title_font}}{Font family for group of procedures title}
  }}
}
\value{
dot graph as string
}
\description{
The visualization itself requires some dependencies to be installed (e.g., \code{DiagrammeR} package). If you do not have those you can \code{cat()} the returned string to the R console and copy it to some web tool for dot visualization (e.g.http://magjac.com/graphviz-visual-editor)
}
