% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{make_dot_nodes}
\alias{make_dot_nodes}
\title{Generates description of dot graph nodes.}
\usage{
make_dot_nodes(
  procedures_list,
  node_prefix = NULL,
  ...,
  procedures_group_bg_color = "#ecf0f1",
  procedures_group_title_font = "Helvetica"
)
}
\arguments{
\item{procedures_list}{List of procedures. See \code{sandardize} for how to specify of procedures list.}

\item{node_prefix}{pefix for node IDs}

\item{...}{
  Arguments passed on to \code{\link[=paste_dot_node]{paste_dot_node}}
  \describe{
    \item{\code{procedure}}{Procedure's function either as charater string or as call (with optional arguments)}
    \item{\code{procedure_title}}{This will overwrites the node's title. By default the title is either from procedure's attribute 'title' (prefixed by \code{procedure_attr_prefix}) or just a name of procedure if there is no such attibute.}
    \item{\code{id}}{Node ID that is used for refering in edges (arrows) definition}
    \item{\code{node_bg_color}}{node's background color}
    \item{\code{node_header_color}}{node's header background color}
    \item{\code{node_font}}{main font family}
    \item{\code{node_fix_width_font}}{fixed with font family}
    \item{\code{add_example}}{wheather to add example from procedure's attribure 'example' (prefixed by \code{procedure_attr_prefix})}
    \item{\code{add_procedure_name}}{wheather to add procedure's name}
    \item{\code{add_procedure_args}}{wheather to add optional procedure's arguments}
    \item{\code{procedure_attr_prefix}}{prefix used to specify procedures' attibutes}
  }}

\item{procedures_group_bg_color}{Background color for visuallizing groupped procedures (nested list of procedures)}

\item{procedures_group_title_font}{Font family for group of procedures title}
}
\value{
dot nodes as string
}
\description{
Generates description of dot graph nodes.
}
