% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_make_alias.r
\name{make_roxy_tags}
\alias{make_roxy_tags}
\title{title
tables (with reference, pp, table no. and title)
example
description (very short and just ref to pp)
reference (url, doi, authors, can I use noweb with papis call to ref to extract all that?)}
\usage{
make_roxy_tags(f)
}
\description{
title
tables (with reference, pp, table no. and title)
example
description (very short and just ref to pp)
reference (url, doi, authors, can I use noweb with papis call to ref to extract all that?)
}
