% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nstandr.r
\name{make_std_names}
\alias{make_std_names}
\alias{standardize}
\title{Makes standard names. Takes either vector or column in the table.}
\usage{
make_std_names(
  x,
  procedures = nstandr_default_procedures_table,
  show_progress = TRUE,
  nrows_min_to_show_progress = 10^5,
  progress_step_nrows = NULL,
  progress_step_in_percent = 1,
  progress_message_use_names = TRUE,
  quite = FALSE,
  save_intermediate_x_to_var = NULL,
  ...
)
}
\arguments{
\item{x}{object (table)}

\item{procedures}{Named list of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments. Names of the list elements are used for progress messages. Procedures can also be passed as data.frame in which case it will be converted to list of procedures with \code{standardize_make_procedures_list} (see its help for the correct format of data.frame with procedures). Default is \code{nstandr_default_procedures_table}}

\item{show_progress}{Whether to report progress percentage. Default is TRUE}

\item{nrows_min_to_show_progress}{The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5}

\item{progress_step_nrows}{If set it will divide the x into chunk of this amount of rows. Default is NULL.}

\item{progress_step_in_percent}{Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.}

\item{progress_message_use_names}{Should we use names from \code{procedures} list to report progress. Default is TRUE.}

\item{quite}{Suppress all messages. Default is FALSE.}

\item{save_intermediate_x_to_var}{For debuging of standartization procedures. Saves intermediate results to this variable. If procedures finish without errors then the variable will be removed.}
}
\value{
standardized data
}
\description{
Makes standard names. Takes either vector or column in the table.
}
