% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nstandr.r
\docType{package}
\name{nstandr-package}
\alias{nstandr}
\alias{nstandr-package}
\title{nstandr: Name Standardization in R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stas Vlasov \email{s.vlasov@uvt.nl}

}
\keyword{internal}
