% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{paste_dot_node}
\alias{paste_dot_node}
\title{Makes a dot node (as html table) from procedure's attributes.}
\usage{
paste_dot_node(
  procedure,
  id,
  procedure_title = NULL,
  node_bg_color = "white",
  node_header_color = "#bdc3c7",
  node_font = "Helvetica",
  node_fix_width_font = "Courier",
  add_example = TRUE,
  add_procedure_name = FALSE,
  add_procedure_args = FALSE,
  procedure_attr_prefix = "nstandr_procedure_"
)
}
\arguments{
\item{procedure}{Procedure's function either as charater string or as call (with optional arguments)}

\item{id}{Node ID that is used for refering in edges (arrows) definition}

\item{procedure_title}{This will overwrites the node's title. By default the title is either from procedure's attribute 'title' (prefixed by \code{procedure_attr_prefix}) or just a name of procedure if there is no such attibute.}

\item{node_bg_color}{node's background color}

\item{node_header_color}{node's header background color}

\item{node_font}{main font family}

\item{node_fix_width_font}{fixed with font family}

\item{add_example}{wheather to add example from procedure's attribure 'example' (prefixed by \code{procedure_attr_prefix})}

\item{add_procedure_name}{wheather to add procedure's name}

\item{add_procedure_args}{wheather to add optional procedure's arguments}

\item{procedure_attr_prefix}{prefix used to specify procedures' attibutes}
}
\value{
dot node as string
}
\description{
Makes a dot node (as html table) from procedure's attributes.
}
