% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{paste_dot_node_tr_td}
\alias{paste_dot_node_tr_td}
\title{Makes TR TD record for dot node TABLE}
\usage{
paste_dot_node_tr_td(str, bg_color = NULL, font = NULL, b = FALSE, i = FALSE)
}
\arguments{
\item{str}{content}

\item{bg_color}{row's background color}

\item{font}{row's font family}

\item{b}{make it bold}

\item{i}{make it intallic}
}
\value{
string
}
\description{
Makes TR TD record for dot node TABLE
}
