% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{save_dot_graph_as}
\alias{save_dot_graph_as}
\title{Saves dot graph as file using system command 'dot' from GraphViz (https://graphviz.org/) if installed.}
\usage{
save_dot_graph_as(
  dot_graph,
  save_to_file = NULL,
  save_as = c("png", "svg", "pdf"),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{dot_graph}{Dot graph as a character string}

\item{save_to_file}{Name of file to save to. Without path or relative path saves to the R's current working directory. If NULL returns file content as string.}

\item{save_as}{File format. Default is "png"}

\item{width}{width of the procedures graph in pixels}

\item{height}{height of the procedures graph in pixels}
}
\value{
Ethier raw file as character string if \code{save_to_file} is NULL or a file name, i.e. the value of \code{save_to_file} is it is not NULL
}
\description{
Saves dot graph as file using system command 'dot' from GraphViz (https://graphviz.org/) if installed.
}
