% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{standardize.nber.removals}
\alias{standardize.nber.removals}
\title{it assumes that there is no removals for the first word}
\usage{
standardize.nber.removals(org.names)
}
\description{
it assumes that there is no removals for the first word
}
