% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nstandr.r
\name{standardize}
\alias{standardize}
\alias{make_standard_names}
\alias{make_std_names}
\alias{nstand}
\title{Standardizes organizational names. Takes either vector or column in the table.}
\usage{
standardize(
  x,
  procedures = nstandr_default_procedures_list,
  show_progress = TRUE,
  nrows_min_to_show_progress = 10^3,
  progress_step_nrows = NULL,
  progress_step_in_percent = 1,
  progress_message_use_names = TRUE,
  quite = FALSE,
  save_intermediate_x_to_var = NULL,
  ...
)

make_std_names(
  x,
  procedures = nstandr_default_procedures_list,
  show_progress = TRUE,
  nrows_min_to_show_progress = 10^3,
  progress_step_nrows = NULL,
  progress_step_in_percent = 1,
  progress_message_use_names = TRUE,
  quite = FALSE,
  save_intermediate_x_to_var = NULL,
  ...
)

make_standard_names(
  x,
  procedures = nstandr_default_procedures_list,
  show_progress = TRUE,
  nrows_min_to_show_progress = 10^3,
  progress_step_nrows = NULL,
  progress_step_in_percent = 1,
  progress_message_use_names = TRUE,
  quite = FALSE,
  save_intermediate_x_to_var = NULL,
  ...
)

nstand(
  x,
  procedures = nstandr_default_procedures_list,
  show_progress = TRUE,
  nrows_min_to_show_progress = 10^3,
  progress_step_nrows = NULL,
  progress_step_in_percent = 1,
  progress_message_use_names = TRUE,
  quite = FALSE,
  save_intermediate_x_to_var = NULL,
  ...
)
}
\arguments{
\item{x}{object (table)}

\item{procedures}{The procedures that basically comprise the standardization algorithm are specified as a list of either (1) names of procedures function as character strings or as (2) calls where you can provide optional arguments or (3) nested lists that allow user to group procedures. Nesting lists of procedures has an effect on standardization progress reporting and on visualizing algorithms with \code{nstandr_plot}. Technically nested lists are equivalent to plain list of procedures as it should produce same results. Names of the list elements are used for progress messages. For unnamed elements the name of procedure's function will be used for standardization progress reporting. Default is \code{nstandr:::nstandr_default_procedures_list}.}

\item{show_progress}{Whether to report progress percentage. Default is TRUE}

\item{nrows_min_to_show_progress}{The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5}

\item{progress_step_nrows}{If set it will divide the x into chunk of this amount of rows. Default is NULL.}

\item{progress_step_in_percent}{Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.}

\item{progress_message_use_names}{Should we use names from \code{procedures} list to report progress. Default is TRUE.}

\item{quite}{Suppress all messages. Default is FALSE.}

\item{save_intermediate_x_to_var}{For debuging of standartization procedures. Saves intermediate results to this variable. If procedures finish without errors then the variable will be removed.}

\item{...}{
  Arguments passed on to \code{\link[=standardize_options]{standardize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to standardize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (standardized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{x_atomic_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_col_name}}{Use this name for the column with results (standardized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}
}
\value{
standardized data
}
\description{
Standardizes organizational names. Takes either vector or column in the table.
}
