% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_is_x_empty.r
\name{standardize_is_data_empty}
\alias{standardize_is_data_empty}
\title{Checks if all elements in vercor(s) are either "", NA, NULL or have zero length}
\usage{
standardize_is_data_empty(x, return_as_true_if_x_zero_length = FALSE)
}
\arguments{
\item{x}{input data to check each vector}

\item{return_as_true_if_x_zero_length}{how to interpret zero lenth input. If TRUE then it returns TRUE. Otherwise NULL.}
}
\value{
logical vector of the same length
}
\description{
Checks if all elements in vercor(s) are either "", NA, NULL or have zero length
}
