% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_is_x_empty.r
\name{standardize_omit_empty}
\alias{standardize_omit_empty}
\title{Removes elements that are either "", NA, NULL or have zero length}
\usage{
standardize_omit_empty(x)
}
\arguments{
\item{x}{vector}
}
\value{
updated vector with empty elements removed
}
\description{
Removes elements that are either "", NA, NULL or have zero length
}
