% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nstandr.r
\name{standardize_x_split}
\alias{standardize_x_split}
\title{Splits the object (table) in chunks by rows}
\usage{
standardize_x_split(x, by, len = NULL)
}
\arguments{
\item{x}{object or table}

\item{by}{number of rows to split by}

\item{len}{length of the table (nrow). If it is NULL then use x_length(x)}
}
\value{
List of (sub)tables
}
\description{
Convenient to apply some function to the table in chunks, e.g., if you want to add display of progress.
}
