% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_unlist_if_possible.r
\name{unlist_if_possible}
\alias{unlist_if_possible}
\title{If column in the \code{x} table is list unlist it if possible}
\usage{
unlist_if_possible(
  x,
  replace_zero_length_with = NA_character_,
  remove_empty_values = TRUE
)
}
\arguments{
\item{x}{object}

\item{replace_zero_length_with}{Default is replace NULLs with NA_character_ because vector of just NA is a logical class}

\item{remove_empty_values}{remove NA, "", etc. from list elements.  (see \code{\link[=standardize_omit_empty]{standardize_omit_empty()}})}
}
\value{
updated object
}
\description{
If column in the \code{x} table is list unlist it if possible
}
