% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{visualize}
\alias{visualize}
\title{Visualizes list of procedures.}
\usage{
visualize(
  procedures_list,
  save_to_file = NULL,
  save_as = NULL,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{procedures_list}{List of procedures. See \code{sandardize} for how to specify of procedures list. Or it can be a character string for standard procedure list (e.g., 'eee-ppat' for Magerman et al., 2006 procedures).}

\item{save_to_file}{If set tries to save visualization to file in the format \code{save_as} with \code{dot} system command (should be installed on your system). If it is NULL tries to open the visualization in browser (this requires DiagrammeR R package to be installed.)}

\item{save_as}{Type of file to save if \code{save_to_file} is not NULL. If not specified (i.e. NULL) then it tries to guess the type of file from \code{save_to_file} extention. Currently "png", "svg" and "pdf" are supported.}

\item{width}{width of the procedures graph in pixels}

\item{height}{height of the procedures graph in pixels}

\item{...}{
  Arguments passed on to \code{\link[=make_dot_graph]{make_dot_graph}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An object of class ‘htmlwidget’ that can be used in Rmarkdown or a file name, i.e. the value of \code{save_to_file} is it is not NULL.
}
\description{
Visualizes list of procedures.
}
