% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_x_length_and_width.r
\name{x_length}
\alias{x_length}
\title{Gets lengths of the object}
\usage{
x_length(x)
}
\arguments{
\item{x}{input data (table)}
}
\value{
Length (\code{nrow}) of the object. If it is atomic it returns its length.
}
\description{
Gets lengths of the object
}
