% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_x_length_and_width.r
\name{x_width}
\alias{x_width}
\title{Gets width of the object}
\usage{
x_width(x)
}
\arguments{
\item{x}{object (table)}
}
\value{
Width (ncol) of the object. If it is atomic it is 1.
}
\description{
Gets width of the object
}
