% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_check_standardize_options.r
\name{check_rows}
\alias{check_rows}
\title{Assumes that rows (if logical) are same length as x}
\usage{
check_rows(rows, x, which_call_to_report = -1L, null_ok = TRUE, na_ok = FALSE)
}
\arguments{
\item{rows}{either numeric of logical vector}

\item{x}{The object to standardize}

\item{which_call_to_report}{When reporting issues which function call to indicate for reference}

\item{null_ok}{Whether NULL is valid value}

\item{na_ok}{Whether NA is valid value}
}
\value{
nothing
}
\description{
Assumes that rows (if logical) are same length as x
}
