% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_escape_patterns.r
\name{escape_regex_for_types}
\alias{escape_regex_for_types}
\title{Escapes special for regex characters conditionally}
\usage{
escape_regex_for_types(
  patterns,
  types,
  escape_fixed = FALSE,
  escape_begins = TRUE,
  escape_ends = TRUE,
  escape_all_if_multiple_types = TRUE
)
}
\arguments{
\item{patterns}{character vector}

\item{types}{character vector of the same length as \code{strings} with
instructions whether and how to to escape regex
characters. Options are:}

\item{escape_fixed}{Whether to escape strings for 'fixed' and
'exact' types in case all types are the same
(length(unique(types)) == 1). The "trim_exact" types get
trimed if not escaped. Otherwise if types heterogenious}

\item{escape_begins}{Whether to escape string for 'begins' type}

\item{escape_ends}{Whether to escape string for 'ends' type}

\item{escape_all_if_multiple_types}{If types are multiple this applies to all esapes options}
}
\value{
string with all special to regex characters escaped
}
\description{
Escapes special for regex characters conditionally

\itemize{
\item fixed  -  Match pattern string as it is within the target vector
\item begins  -  Match pattern string as it is in the beggining of the target vector
\item trim_begins  -  Match pattern string as it is in the beginning of the target vector ignoring preceding white-spaces
\item ends  -  Match pattern string as it is in the end of the target vector
\item trim_ends  -  Match pattern string as it is in the end of the target vector ignoring leading white-spaces
\item exact  -  Match pattern string exactly (i.e., match equal strings)
\item trim_exact  -  Match pattern string exactly (i.e., match equal strings) ignoring surrounding white-spaces
\item regex  -  Match regex pattern
}
}
