% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vector.r
\name{get_vector}
\alias{get_vector}
\title{Gets vector by column and defactor if needed. Optionaly one can provide a fallback_value which will be returned if col is not specified.}
\usage{
get_vector(
  x,
  col,
  rows = NULL,
  choices = NULL,
  fallback_value = NULL,
  fallback_value_ignored_if_col = TRUE,
  fallback_value_any_missing = TRUE,
  fallback_value_ensure_length = TRUE,
  check_x_col_rows = TRUE,
  which_call_to_report = -1L
)
}
\arguments{
\item{x}{Input data. Can be vector, data.frame or a data.table}

\item{col}{Column of interest in the input data \code{x}. The vector we would like to work with. Ignored if input \code{x} is a atomic vector.}

\item{rows}{Rows of interest}

\item{choices}{Optional set of choices that return value should be subset of. Do no check if it is NULL.}

\item{fallback_value}{If col is NULL then return this value (but see \code{fallback_value_ignored_if_col} argument). Also check if it is a character vector and same lenght as x. If it is of length 1 then replicate it to match x's length.}

\item{fallback_value_ignored_if_col}{A bolean toggle. When set if col is provided then fallback_value_is_simply ignored. Otherwise col is ignored when fallback_value is provided.}

\item{fallback_value_any_missing}{Toggle check if missing values are allowed in fallback_value}

\item{fallback_value_ensure_length}{Toggle if fallback_value should we make it same length as \code{x}}

\item{check_x_col_rows}{Toggle wheather to use check_x, check_col, check_rows?}

\item{which_call_to_report}{Which call to report if argument checks fail.}
}
\value{
A vector. Factors in imput \code{data} are converted to string.
}
\description{
Gets vector by column and defactor if needed. Optionaly one can provide a fallback_value which will be returned if col is not specified.
}
