% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.defactor}
\alias{harmonize.defactor}
\title{Defactor the object}
\usage{
harmonize.defactor(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{Arguments passed on to \code{harmonize.defactor.vector}
\describe{
  \item{x}{a vector}
  \item{check.numeric}{check if vector is numeric. Default is TRUE. Takes longer with this check but avoids type conversion (numeric to character).}
}}
}
\value{
object of the same type without factors
}
\description{
Returns object of the same type without factors
}
