% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize}
\alias{harmonize}
\title{Harmonizes organizational names. Takes either vector or column in the table.}
\usage{
harmonize(
  x,
  procedures = harmonize.default.procedures,
  progress = TRUE,
  progress.min = 10^5,
  progress.by = NA,
  progress.percent = 1,
  progress.message.use.names = TRUE,
  quite = FALSE,
  ...
)
}
\arguments{
\item{x}{object (table)}

\item{procedures}{List of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments.}

\item{progress}{Show the progress? Default is TRUE}

\item{progress.min}{The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5}

\item{progress.by}{If set it will divide the x into chunk of this amount of rows. Default is NA.}

\item{progress.percent}{Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.}

\item{progress.message.use.names}{Should we use names from \code{procedures} list to report progress. Default is TRUE.}

\item{quite}{Suppress all messages. Default is FALSE.}

\item{...}{
  Arguments passed on to \code{\link[=harmonize.x]{harmonize.x}}
  \describe{
    \item{\code{inset.vector}}{a vector to inset. Optional. Default is NULL}
    \item{\code{x.col}}{vector of interest in \code{x} object}
    \item{\code{x.col.update}}{Update values in \code{x.col} column. Default is FALSE. If set \code{inset.append}, \code{inset.name} and \code{inset.suffix} are ignored. Also if set the default for \code{return.x.cols.all} will be set to TRUE.}
    \item{\code{x.rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records}
    \item{\code{x.atomic.name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{x.col} will be used.}
    \item{\code{inset.omitted.val}}{If \code{x.rows} is set merge these values to the results. It should be a vector of length 1 or \code{nrow(x)}. If the value is NULL (default) then use values of \code{x.col}.}
    \item{\code{inset.append}}{If set then put \code{inset.vector} as the last instead of first vector/column. Default is FALSE.}
    \item{\code{inset.suffix}}{If \code{inset.name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in \code{return.x.cols} it will add counter at the end to avoid variables with the same names.}
    \item{\code{inset.name}}{Use this name for the first column in results (harmonized names). Default is NULL, which means that either x.atomic.name if x is vector or original x.col name will be used with \code{inset.suffix} at the end.}
    \item{\code{return.x.cols}}{If x is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (x.col).}
    \item{\code{return.x.cols.all}}{Whether to bind all columns in x. Defaults depends on values of \code{x.col.update} and \code{inset.append}. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the return.x.cols is ignored.}
  }}
}
\value{

}
\description{
Harmonizes organizational names. Takes either vector or column in the table.
}
