% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize}
\alias{harmonize}
\title{Harmonizes organizational names. Takes either vector or column in the table.}
\usage{
harmonize(
  x,
  procedures = harmonizer_default_procedures_table,
  show_progress = TRUE,
  nrows_min_to_show_progress = 10^5,
  progress_step_nrows = NULL,
  progress_step_in_percent = 1,
  progress_message_use_names = TRUE,
  quite = FALSE,
  save_intermediate_x_to_var = NULL,
  ...
)
}
\arguments{
\item{x}{object (table)}

\item{procedures}{Named list of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments. Names of the list elements are used for progress messages. Procedures can also be passed as data.frame in which case it will be converted to list of procedures with \code{harmonize_make_procedures_list} (see its help for the correct format of data.frame with procedures). Default is \code{harmonizer_default_procedures_table}}

\item{show_progress}{}

\item{nrows_min_to_show_progress}{The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5}

\item{progress_step_nrows}{If set it will divide the x into chunk of this amount of rows. Default is NULL.}

\item{progress_step_in_percent}{Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.}

\item{progress_message_use_names}{Should we use names from \code{procedures} list to report progress. Default is TRUE.}

\item{quite}{Suppress all messages. Default is FALSE.}

\item{save_intermediate_x_to_var}{For debuging of standartization procedures. Saves intermediate results to this variable. If procedures finish without errors then the variable will be removed.}

\item{...}{
  Arguments passed on to \code{\link[=harmonize_options]{harmonize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to harmonize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (harmonized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{atomic_x_col_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_name}}{Use this name for the column with results (harmonized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}

\item{progress}{Show the progress? Default is TRUE}
}
\value{

}
\description{
Harmonizes organizational names. Takes either vector or column in the table.
}
