% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{harmonize.toupper}
\alias{harmonize.toupper}
\title{Uppercases vector of interest in the object (table)}
\usage{
harmonize.toupper(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{
  Arguments passed on to \code{\link[=inset_vector]{inset_vector}}
  \describe{
    \item{\code{vector}}{Character vector to inset into the \code{x} object}
    \item{\code{x}}{Data to harmonize. Character vector or data.frame or data.table}
  }}
}
\value{
updated data (as data.table)
}
\description{
Uppercases vector of interest in the object (table)
}
