% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_options.r
\name{harmonize_options}
\alias{harmonize_options}
\title{Does nothing but stores (as its own default arguments) options that control vector handeling through harmonization process. These options are available in most harmonizer functions that accept \code{...} parameter.}
\usage{
harmonize_options(
  col = 1,
  rows = NULL,
  omitted_rows_values = NULL,
  output = "replace_col",
  name = NULL,
  name_for_x_atomic = "x",
  name_suffix = "_harmonized",
  append_copy = FALSE,
  append_copy_name_format = "\%name_harmonizing_\%number_\%procedure",
  select_cols = NULL
)
}
\arguments{
\item{col}{Column of interest (the one we need to harmonize) in the \code{x} object (if it is data.frame like).}

\item{rows}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}

\item{omitted_rows_values}{If \code{rows} parameter is set then merge \code{omitted_rows_values} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}

\item{output}{Where to inset retults (harmonized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}

\item{name}{Use this name for the column with results (harmonized values). Default is NA, which means that either \code{name_for_x_atomic} if \code{x} is vector or original col name will be used with \code{name_suffix} at the end.}

\item{name_for_x_atomic}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}

\item{name_suffix}{If \code{name} is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in \code{select_x_cols} it will add counter at the end to avoid variables with the same names.}

\item{append_copy}{Whether to append a copy of result vector to \code{x} object}

\item{append_copy_name_format}{How the append copy wiil be named}

\item{select_cols}{If x` object is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (col).}
}
\value{
Always NULL. It does nothing.
}
\description{
Does nothing but stores (as its own default arguments) options that control vector handeling through harmonization process. These options are available in most harmonizer functions that accept \code{...} parameter.
}
