% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\docType{package}
\name{harmonizer-package}
\alias{harmonizer}
\alias{harmonizer-package}
\title{harmonizer: Harmonization of Organizational Names}
\description{
Harmonizes organizational names using steps described in Thoma et al. (2010) and Magerman, Looy, Bart, & Song (2006) and more.
}
\details{
Harmonizer package standardizes (harmonizes) organizational names
mainly using procedures described in Thoma et al. (2010) and
Magerman, Looy, Bart, & Song (2006) but not only.  This is work
in progress. Please, file an issues or suggestion if you have
any.  The main function is \code{\link[=harmonize]{harmonize()}}.
}
\author{
\strong{Maintainer}: Stanislav Vlasov \email{s.vlasov@uvt.nl}

}
\keyword{internal}
