% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handel_target.r
\name{inset_target}
\alias{inset_target}
\title{Insets target vector back to input object (\code{x})}
\usage{
inset_target(
  vector,
  x,
  omitted_rows_value_for_new_col = NULL,
  allow_na_in_vector = TRUE,
  which_call_to_report = -5L,
  return_only_target_col = FALSE,
  ...
)
}
\arguments{
\item{vector}{Character vector to inset into the \code{x} object}

\item{x}{Data to harmonize. Character vector or data.frame or
data.table}

\item{omitted_rows_value_for_new_col}{Alternative value
\code{omitted_rows_value} to use in case we create new column in
x. For example, it is use in insetting codes to avoid the
default \code{omitted_rows_value} use initial \code{col} in which case
codes will be mixed with input values}

\item{allow_na_in_vector}{Whether to allow NA in inset vector}

\item{which_call_to_report}{System call number (e.g., -2L) to
include in report if arguments checks fails}

\item{return_only_target_col}{If toggled to TRUE then only return
the vector to be inset (output_placement argument is ignored)}

\item{...}{
  Arguments passed on to \code{\link[=harmonize_options]{harmonize_options}}
  \describe{
    \item{\code{col}}{Column of interest (the one we need to harmonize) in the \code{x} object (if it is data.frame like).}
    \item{\code{rows}}{Logical vector to filter records of interest. Default is NULL which means do not filter records.}
    \item{\code{omitted_rows_value}}{If \code{rows} parameter is set then merge \code{omitted_rows_value} with the results (filtered by \code{rows}). Either single string or a character vector of length \code{nrow(x)}. If NULL (the default) then original values of \code{col} are merged with results.}
    \item{\code{output_placement}}{Where to inset retults (harmonized vector) in the \code{x} object. Default options is 'replace_col' which overwrides the \code{col} in \code{x} with results. Other options:
\itemize{
\item 'omit' :: do not write results back to table (usually used when \code{append_output_copy} is set for temporary values)
\item 'prepend_to_col' :: prepend to \code{col}
\item 'append_to_col' :: append to \code{col}
\item 'prepend_to_x' :: prepend to \code{x} data.frame like object
\item 'append_to_x' :: append to \code{x} data.frame like object
}}
    \item{\code{atomic_x_col_name}}{If \code{x} is vector use this name for original column if it is in results. Default is "x". If \code{x} is table the name of \code{col} will be used.}
    \item{\code{output_name}}{Use this name for the column with results (harmonized values). Parts in curly brakeds are substitute strings. Options for substitutions are:}
    \item{\code{append_output_copy}}{Whether to append a copy of result vector to \code{x} object}
    \item{\code{output_copy_col_name}}{How the append copy wiil be named}
  }}
}
\value{
Data.table or character vector
}
\description{
Insets target vector back to input object (\code{x})
}
