% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vector_inset_vector.r
\name{inset_vector}
\alias{inset_vector}
\title{Insets target vector back to input object (\code{data})}
\usage{
inset_vector(vector, x, ...)
}
\description{
Insets target vector back to input object (\code{data})
}
