% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magerman.r
\name{magerman.detect.legal.form}
\alias{magerman.detect.legal.form}
\title{Detects legal form}
\usage{
magerman.detect.legal.form(x, ...)
}
\arguments{
\item{x}{table}

\item{...}{
  Arguments passed on to \code{\link[=magerman.detect.legal.form.end]{magerman.detect.legal.form.end}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Harmonized table
}
\description{
Detects legal form
}
