% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizer.r
\name{unlist_if_possible}
\alias{unlist_if_possible}
\title{If column in the \code{x} table is list unlist it if possible}
\usage{
unlist_if_possible(x)
}
\arguments{
\item{x}{object}
}
\value{
updated object
}
\description{
If column in the \code{x} table is list unlist it if possible
}
